+++ .circleci/build-and-test/jobs.yml
@@ -57,9 +57,6 @@
           name: Set up cypress test users
           command: cd tdrs-backend; docker-compose exec web python manage.py generate_cypress_users
       - run:
-          name: Set up cypress test super user
-          command: cd tdrs-backend; docker-compose exec web python manage.py generate_cypress_superuser
-      - run:
           name: Run Cypress e2e tests
           command: cd tdrs-frontend; npm run test:e2e-ci
       - store_artifacts:
