+++ tdrs-frontend/cypress.config.js
@@ -6,15 +6,10 @@ module.exports = defineConfig({
   e2e: {
     baseUrl: 'http://localhost:3000',
     specPattern: '**/*.feature',
-    experimentalSessionAndOrigin: true,
 
     env: {
       apiUrl: 'http://localhost:8080/v1',
       cypressToken: 'local-cypress-token',
-      cypressName: 'cypress',
-      cypressStt: '6',
-      cypressSttName: 'Colorado',
-      cypressGroup: '2',
     },
 
     async setupNodeEvents(on, config) {
