+++ tdrs-frontend/cypress/e2e/common-steps/common-steps.js
@@ -5,14 +5,18 @@ When('The user visits the home page', () => {
   cy.contains('Sign into TANF Data Portal', { timeout: 30000 })
 })
 
-When('The new user logs in', () => {
-  cy.login('new-cypress@teamraft.com')
-})
-
 When('The admin logs in', () => {
   cy.adminLogin('cypress-admin@teamraft.com')
 })
 
-When('The user logs in', () => {
-  cy.login('new-cypress@teamraft.com')
+When('{string} logs in', (username) => {
+  cy.login(username)
+})
+
+When('The admin sets the approval status of {string} to {string}', (username, status) => {
+  cy.get('@cypressUser').then((cypressUser) => {
+    cy.get('@adminCsrfToken').then((csrfToken) => {
+      cy.changeUserInfo(`${cypressUser.selector.id}`, `${csrfToken}`, username, '6', '2', status)
+    })
+  })
 })
