+++ tdrs-frontend/cypress/support/commands.js
@@ -100,14 +100,14 @@ Cypress.Commands.add('adminApiRequest', (options = {}) => {
   }
 })
 
-Cypress.Commands.add('reinitUserAccount', (user, token) => {
+Cypress.Commands.add('reinitUserAccount', (userID, token) => {
   let options = {
     method: 'POST',
-    url: `http://localhost:8080/admin/users/user/${user}/change/`,
+    url: `http://localhost:8080/admin/users/user/${userID}/change/`,
     headers: { X_CSRFTOKEN: token },
     form: true,
     body: {
-      username: 'new-cypress@teamraft.com', // This `'${cypressUser.selector.username}'` should work but it does not. There are extra characters in the username encoding for some reason.
+      username: 'new-cypress@teamraft.com',
       first_name: '',
       last_name: '',
       email: 'new-cypress@teamraft.com',
@@ -119,17 +119,15 @@ Cypress.Commands.add('reinitUserAccount', (user, token) => {
   cy.adminApiRequest(options)
 })
 
-Cypress.Commands.add('changeUserInfo', (user, token, fname, lname, stt, group, status) => {
+Cypress.Commands.add('changeUserInfo', (userID, token, username, stt, group, status) => {
   let options = {
     method: 'POST',
-    url: `http://localhost:8080/admin/users/user/${user}/change/`,
+    url: `http://localhost:8080/admin/users/user/${userID}/change/`,
     headers: { X_CSRFTOKEN: token },
     form: true,
     body: {
-      username: 'new-cypress@teamraft.com', // This `'${cypressUser.selector.username}'` should work but it does not. There are extra characters in the username encoding for some reason.
-      first_name: fname,
-      last_name: lname,
-      email: 'new-cypress@teamraft.com',
+      username: username,
+      email: username,
       stt: stt,
       groups: group,
       account_approval_status: status,
